chore: code optimize
